refactor(formatter): implement `Format::fmt` directly for the enum AST instead of delegating to `FormatWrite::write` (#13314)
Note: I haven't started to refactor the generator files yet. I just found that implementing `FormatWrite::write` for enum AST is unnecessary, so I made it clearer so that I can develop the remaining features.
