+++ tasks/ast_tools/src/generators/formatter/mod.rs
@@ -1,7 +1,5 @@
 mod ast_nodes;
 mod format;
-mod format_write;
 
 pub use ast_nodes::FormatterAstNodesGenerator;
 pub use format::FormatterFormatGenerator;
-pub use format_write::FormatterFormatWriteGenerator;
