+++ tasks/ast_tools/src/generators/mod.rs
@@ -19,9 +19,7 @@ mod visit;
 pub use assert_layouts::AssertLayouts;
 pub use ast_builder::AstBuilderGenerator;
 pub use ast_kind::AstKindGenerator;
-pub use formatter::{
-    FormatterAstNodesGenerator, FormatterFormatGenerator, FormatterFormatWriteGenerator,
-};
+pub use formatter::{FormatterAstNodesGenerator, FormatterFormatGenerator};
 pub use get_id::GetIdGenerator;
 pub use raw_transfer::RawTransferGenerator;
 pub use raw_transfer_lazy::RawTransferLazyGenerator;
