+++ tasks/ast_tools/src/main.rs
@@ -288,7 +288,6 @@ const GENERATORS: &[&(dyn Generator + Sync)] = &[
     &generators::TypescriptGenerator,
     &generators::FormatterFormatGenerator,
     &generators::FormatterAstNodesGenerator,
-    &generators::FormatterFormatWriteGenerator,
 ];
 
 /// Attributes on structs and enums (not including those defined by derives/generators)
